Move legacy editor to respective folder
