+++ app/[user]/draft/[id]/page.tsx
@@ -1,12 +1,5 @@
-'use client'
-
-import { Editor } from '@/components/editor'
-import { Preview } from '@/components/preview'
-import { Separator } from '@/components/ui/separator'
-import { useMediaquery } from '@/hooks/use-media-query'
-import { useSyncScroll } from '@/hooks/use-sync-scroll'
+import LegacyEditor from '@/components/legacy-editor'
 import { getDrafts } from '@/lib/supabase'
-import { useRef, useState } from 'react'
 
 export async function generateStaticParams() {
 	const { data } = await getDrafts()
@@ -21,17 +14,9 @@ export async function generateStaticParams() {
 }
 
 export default function DraftsPage() {
-	const editorRef = useRef<HTMLDivElement>(null)
-	const previewRef = useRef<HTMLDivElement>(null)
-	const [content, setContent] = useState<string>('')
-	const isDesktop = useMediaquery('(min-width: 768px)')
-
-	useSyncScroll(editorRef, previewRef)
 	return (
 		<main className="container flex h-[calc(100%-4rem)] flex-col md:flex-row xl:border-x">
-			<Editor editorRef={editorRef} content={content} setContent={setContent} />
-			<Separator orientation={isDesktop ? 'vertical' : 'horizontal'} />
-			<Preview previewRef={previewRef} content={content} />
+			<LegacyEditor />
 		</main>
 	)
 }
